Start db refactoring
