+++ tests/conftest.py
@@ -17,7 +17,7 @@ from udata_hydra import config
 from udata_hydra.app import app_factory
 import udata_hydra.cli  # noqa - this register the cli cmds
 from udata_hydra.logger import stop_sentry
-from udata_hydra.utils.db import insert_check, update_check
+from udata_hydra.db import checks
 
 DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://postgres:postgres@localhost:5433/postgres")
 RESOURCE_ID = "c4e3a9fb-4415-488e-ba57-d05269b27adf"
@@ -179,10 +179,10 @@ async def fake_check():
             "detected_last_modified_at": detected_last_modified_at,
             "parsing_table": hashlib.md5(url.encode("utf-8")).hexdigest() if parsing_table else None,
         }
-        id = await insert_check(data)
+        id = await checks.insert(data)
         data["id"] = id
         if created_at:
-            await update_check(id, {"created_at": created_at})
+            await checks.update(id, {"created_at": created_at})
             data["created_at"] = created_at
         return data
