+++ udata_hydra/app.py
@@ -9,6 +9,7 @@ from marshmallow import Schema, fields, ValidationError
 
 from udata_hydra import context, config
 from udata_hydra.crawl import get_excluded_clause
+from udata_hydra.db import catalog
 from udata_hydra.logger import setup_logging
 from udata_hydra.utils.minio import delete_resource_from_minio
 from udata_hydra.worker import QUEUES
@@ -88,14 +89,13 @@ async def resource_created(request):
     dataset_id = valid_payload["dataset_id"]
     resource_id = valid_payload["resource_id"]
 
-    pool = request.app["pool"]
-    async with pool.acquire() as connection:
-        # Insert new resource in catalog table and mark as high priority for crawling
-        q = f"""
-                INSERT INTO catalog (dataset_id, resource_id, url, deleted, priority)
-                VALUES ('{dataset_id}', '{resource_id}', '{resource["url"]}', FALSE, TRUE)
-                ON CONFLICT (dataset_id, resource_id, url) DO UPDATE SET priority = TRUE;"""
-        await connection.execute(q)
+    # Insert new resource in catalog table and mark as high priority for crawling
+    await catalog.insert(
+        ("dataset_id", "resource_id", "url", "deleted", "priority"),
+        (dataset_id, resource_id, resource["url"], False, True),
+        on_conflict="UPDATE SET priority = TRUE",
+        pool=request.app["pool"],
+    )
 
     return web.json_response({"message": "created"})
 
@@ -124,12 +124,15 @@ async def resource_updated(request):
         if len(res):
             q = f"""UPDATE catalog SET priority = TRUE, url = '{resource["url"]}'
             WHERE resource_id = '{resource_id}';"""
+            await connection.execute(q)
         else:
-            q = f"""
-                    INSERT INTO catalog (dataset_id, resource_id, url, deleted, priority)
-                    VALUES ('{dataset_id}', '{resource_id}', '{resource["url"]}', FALSE, TRUE)
-                    ON CONFLICT (dataset_id, resource_id, url) DO UPDATE SET priority = TRUE;"""
-        await connection.execute(q)
+            # Insert new resource in catalog table and mark as high priority for crawling
+            await catalog.insert(
+                ("dataset_id", "resource_id", "url", "deleted", "priority"),
+                (dataset_id, resource_id, resource["url"], False, True),
+                on_conflict="UPDATE SET priority = TRUE",
+                pool=request.app["pool"],
+            )
 
     return web.json_response({"message": "updated"})
